UC01-NominalEntry runs successfully with Ciera-generated code
Ciera-generated code produces an exception
(https://github.com/johnrwolfe/CarPark/issues/32), so this commit
comments out the offending actions, which are not required just now.
Also moved the configuration project up to the root of the
InteractiveTesting project because Ciera seems unable to find it if it
is contained in another package.