Updated tests
Presently the updated support for Time is not working as I expect.  This
may be a misunderstanding on my part or a defect.
Also, to run these tests in Verifier requires the removal of the Ciera
runtime project from the workspace and the addition of the EEs from the
Epoch test case.