Update tests and interfaces to support them
Replaced TestSupport interface with four messages in OperatorConsole
(OpenCarpark, CarparkOpened, CloseCarpark, CarparkClosed) to provide a
closed-loop protocol between the testbench and the CarparkControl
component to ensure that initialization of the carpark is complete
before the testbench begins executing test cases.
Updated NominalBucket to use the new messages.
This commit has not yet been tested.