Client to server message works
It seems that JSON field names must be all lower case?  When
VehicleWaitingMsg contained a member called "Location", the message was
being sent, but the payload was null.  Changing the field name in the
html and JavaScript as well as the member name in the POJ message class
to "location" solved the problem.