Added barrier state to entry stand client
Notice that this commit leaves the "VehicleArrived" button active all
the time.  This allows testing of the race condition between one vehicle
entering the car park and triggering the "VehicleEntered" sensor and
another vehicle behind the first, rolling forward to the entry stand and
triggering the "VehicleArrived" sensor.  Refer to #45 for details.