Tested successfully with Verifier.
Slight modifications to TimeUtilities, which unfortunately must be
maintained in each component that needs it because
https://support.onefact.net/issues/11958.
Also, while it is tempting to connect the TestSupport component to the
InteractiveTestbench component so the latter may rely on the former to
manipulate the time base (thereby maintaining a single copy of
TimeUtilities residing only in the TestSupport component), it does not
quite work.  Since Ciera does not support synchronous interface
operations, the TestSupport component must surface an interface composed
of asynchronous signals.  This leaves the TestBench without the ability
to know when the time base has been adjusted so that it can proceed with
the next step in the test.  Working around the lack of synchronous
interface operations requires additional (read noisy) events and
transitions to the NominalBucket state model.  Rather than add that
noise to the model, we will maintain two copies of TimeUtilities so the
TestBench can invoke them synchronously.