Complete
The TestControl client allows the user to set the current time to a
specified date and time or advance the current time by a specified
duration.
The intended use is for test and demonstration.  For example, set the
current time to 7:42 a.m. on the current day, then make a vehicle enter
the carpark, then advance time by five hours, then, prod the patron to
pay for the stay, then advance time by seven minutes, and finally, make
the vehicle exit the carpark.