Complete
Runs successfully on Verifier and Ciera-generated code with
browser-based clients.  Tested UC01, UC05, UC06, and UC07 with the
latter, including a variation on UC07 in which the patron cancels the
transaction when the payment machine has insufficient change and then
takes the same ticket to another payment machine with sufficient change
and completes payment there.
The 'elif' issue mentioned in a pervious commit is
https://support.onefact.net/issues/11978.