Cleaned up several things
After tiring of inadvertently causing can't happen faults by poking
Ticket Inserted before Vehicle Arrived on the Exit Stand, I made the
Exit Stand leave the former disabled until the latter has been pushed.
The ExitStand state machine was not bothering to send the UnpaidStayExit
message to the Operator Console.  This commit rectifies that oversight.
This commit also adds a Duration attribute to Charge because the
UnpaidStayExit message carries it as a parameter so the operator can
make an informed decision to waive (or not) the charge for an unpaid
stay.  Suppose the stay was only a few minutes, but a step-function or
minimum-fee demands a substantial sum.  The duration of the stay is
therefore useful information in addition to the assessed fee.
The difference between Vue's v-if and v-show is still unclear, but it
seems that buttons defined within v-if never do anything when pushed
while those defined within v-show cause an action.
This commit still has a number of issues including a lack of action when
"Cancel Extra" is pushed for an unpaid stay.