Use a day in the future for testing
To work around https://support.onefact.net/issues/11980, this commit
establishes a fee schedule for 01-Apr-2021 and makes the testbench set
the current time to that same day for testing.
With these changes, the timer that provides periodic updates of date and
time to the Operator Console works just fine.
Tested successfully with the nominal test bucket on both Verifier and
Ciera-generated code with the browser-based clients.