Make Deployment the default configuration
Until we learn how to use a single pom.xml file to support multiple
build configurations, we are maintaining one file for each configuration
(e.g., pom-clients.xml, pom-testbench.xml, pom-autotest.xml).
We generally leave pom.xml as a copy of pom-clients.xml so that the
default build is for the Deployment configuration supporting
browser-based clients.